backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
PyPDF2==3.0.1
python-dotenv==1.0.0
pydantic[email]==2.5.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
motor[srv]==3.3.2
